requirements.txt
alembic==1.6.4
aniso8601==9.0.1
attrs==21.2.0
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
Werkzeug==2.0.1
